myblog/requirements.txt
appdirs==1.4.3
branca==0.2.0
Django==1.10.6
folium==0.3.0
googlemaps==2.4.6
Jinja2==2.9.5
MarkupSafe==1.0
nltk==3.2.2
oauthlib==2.0.1
packaging==16.8
pyparsing==2.2.0
python-gmaps==0.3.1
pytz==2016.10
requests==2.11.1
requests-oauthlib==0.8.0
six==1.10.0
textblob==0.12.0
tweepy==3.5.0
